fake_news_chatbot/requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
groq>=0.9.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.2
chromadb>=0.4.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
langdetect>=1.0.9
deep-translator>=1.11.4
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
